* Better language support * New translations only in English and Spanish Details: * Moved Forgot view to Account folder because was not matching with translation folder * Added translations to home page * Added translations to Profile page * Added translations to delete role popup * Added translations to delete confirmation popup and created resource files * Added translations to logout confirmation popup and created resource files * Changed logout icon * Added translations to main layout and created resource files
