+++ BlazorHero.CleanArchitecture/Client/Pages/Account/Forgot.razor
@@ -21,7 +21,7 @@
             </div>
         </MudItem>
         <MudItem xs="12">
-            <MudTextField T="string" Label="@localizer["E-mail"]"  @bind-Value="model.Email" For="@(() => model.Email)" Required="true"  Variant="Variant.Outlined"></MudTextField>
+            <MudTextField T="string" Label="@localizer["E-Mail"]"  @bind-Value="model.Email" For="@(() => model.Email)" Required="true"  Variant="Variant.Outlined"></MudTextField>
         </MudItem>
         <MudItem xs="12" Class="d-flex justify-center">
             <MudButton Variant="Variant.Filled" Color="Color.Primary" Link="/authentication/reset-password" Size="Size.Large" FullWidth="true">@localizer["Reset Password"]</MudButton>
