+++ BlazorHero.CleanArchitecture/Client/Pages/Account/Forgot.razor.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace BlazorHero.CleanArchitecture.Client.Pages.Identity
+namespace BlazorHero.CleanArchitecture.Client.Pages.Account
 {
     public partial class Forgot
     {
