+++ BlazorHero.CleanArchitecture/Client/Pages/Content/Home.razor
@@ -19,8 +19,8 @@
         </MudItem>
         <MudItem xs="12" sm="12" md="12">
             <div class="pa-4 justify-center my-4 mud-text-align-center">
-                <MudButton Variant="Variant.Filled" Color="Color.Primary" Link="">Project Page</MudButton>
-                <MudButton Variant="Variant.Filled" Color="Color.Secondary" Link="https://www.buymeacoffee.com/codewithmukesh" Target="_blank">Support Me!</MudButton>
+                <MudButton Variant="Variant.Filled" Color="Color.Primary" Link="https://github.com/blazorhero/CleanArchitecture">@_localizer["Project Page"]</MudButton>
+                <MudButton Variant="Variant.Filled" Color="Color.Secondary" Link="https://www.buymeacoffee.com/codewithmukesh" Target="_blank">@_localizer["Support Me"]</MudButton>
             </div>
         </MudItem>
     </MudGrid>
