+++ BlazorHero.CleanArchitecture/Client/Resources/Pages/Content/Home.en.resx
@@ -120,6 +120,12 @@
   <data name="Description" xml:space="preserve">
     <value>Clean Architecture Template For Blazor Web-Assembly</value>
   </data>
+  <data name="Project Page" xml:space="preserve">
+    <value>Project Page</value>
+  </data>
+  <data name="Support Me" xml:space="preserve">
+    <value>Support Me!</value>
+  </data>
   <data name="Title" xml:space="preserve">
     <value>Blazor Hero</value>
   </data>
