+++ BlazorHero.CleanArchitecture/Client/Resources/Pages/Identity/Profile.es.resx
@@ -117,6 +117,9 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
+  <data name="Delete Image" xml:space="preserve">
+    <value>Borrar Imagen</value>
+  </data>
   <data name="Delete Picture" xml:space="preserve">
     <value>Borrar Imagen</value>
   </data>
@@ -140,6 +143,9 @@
   </data>
   <data name="Save Changes" xml:space="preserve">
     <value>Guardar cambios</value>
+  </data>
+  <data name="Upload Image" xml:space="preserve">
+    <value>Cargar Imagen</value>
   </data>
   <data name="Your Profile has been updated. Please Login to Continue." xml:space="preserve">
     <value>Tu perfil fue actualizado, por favor ingrese para continuar.</value>
