+++ BlazorHero.CleanArchitecture/Client/Resources/Pages/Identity/Roles.en.resx
@@ -126,6 +126,9 @@
   <data name="Delete" xml:space="preserve">
     <value>Delete</value>
   </data>
+  <data name="Delete Content" xml:space="preserve">
+    <value>Do you want to delete the role with Id {0}?</value>
+  </data>
   <data name="Edit" xml:space="preserve">
     <value>Edit</value>
   </data>
