+++ BlazorHero.CleanArchitecture/Client/Resources/Shared/Dialogs/Logout.es.resx
@@ -117,16 +117,10 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="Description" xml:space="preserve">
-    <value>Clean Architecture Template para Blazor Web-Assembly</value>
+  <data name="Cancel" xml:space="preserve">
+    <value>Cancelar</value>
   </data>
-  <data name="Project Page" xml:space="preserve">
-    <value>Página del proyecto</value>
-  </data>
-  <data name="Support Me" xml:space="preserve">
-    <value>Apoyarme!</value>
-  </data>
-  <data name="Title" xml:space="preserve">
-    <value>Blazor Hero</value>
+  <data name="Logout Confirmation" xml:space="preserve">
+    <value>Confirmación de desconexión</value>
   </data>
 </root>
