+++ BlazorHero.CleanArchitecture/Client/Shared/Dialogs/Logout.razor
@@ -1,10 +1,17 @@
 ﻿@namespace BlazorHero.CleanArchitecture.Client.Shared.Dialogs
+@inject Microsoft.Extensions.Localization.IStringLocalizer<Logout> localizer
 <MudDialog>
+    <TitleContent>
+        <MudText Typo="Typo.h6">
+            <MudIcon Icon="@Icons.Material.Filled.Logout" Class="mr-3 mb-n1" />
+            @localizer["Logout Confirmation"]
+        </MudText>
+    </TitleContent>
     <DialogContent>
         <MudText>@ContentText</MudText>
     </DialogContent>
     <DialogActions>
-        <MudButton OnClick="Cancel">Cancel</MudButton>
+        <MudButton OnClick="Cancel">@localizer["Cancel"]</MudButton>
         <MudButton Color="@Color" Variant="Variant.Filled" OnClick="Submit">@ButtonText</MudButton>
     </DialogActions>
 </MudDialog>
