+++ BlazorHero.CleanArchitecture/Client/Shared/MainLayout.razor
@@ -1,4 +1,5 @@
 ﻿@inherits LayoutComponentBase
+@inject Microsoft.Extensions.Localization.IStringLocalizer<MainLayout> localizer
 
 <MudThemeProvider Theme="currentTheme" />
 <MudDialogProvider />
@@ -23,17 +24,17 @@
                     <MudButton Variant="Variant.Filled" Color="Color.Secondary" Link="https://www.buymeacoffee.com/codewithmukesh" Target="_blank">Buy Me A Coffee!</MudButton>
 
                     <MudDivider Vertical="true" FlexItem="true" DividerType="DividerType.Middle" Class="mx-4 my-5" />
-                    <MudMenu EndIcon="@Icons.Filled.KeyboardArrowDown" Label="Support" Color="Color.Inherit" Dense="true" Direction="Direction.Right" OffsetY="true">
-                        <MudText Typo="Typo.body2" Class="px-4 py-2"><b>Community</b></MudText>
-                        <MudMenuItem Link="https://www.facebook.com/groups/about.asp.net.core" Target="_blank">Facebook Group</MudMenuItem>
-                        <MudMenuItem Link="https://www.facebook.com/codewithmukesh" Target="_blank">Facebook Page</MudMenuItem>
+                    <MudMenu EndIcon="@Icons.Filled.KeyboardArrowDown" Label="@localizer["Support"]" Color="Color.Inherit" Dense="true" Direction="Direction.Right" OffsetY="true">
+                        <MudText Typo="Typo.body2" Class="px-4 py-2"><b>@localizer["Community"]</b></MudText>
+                        <MudMenuItem Link="https://www.facebook.com/groups/about.asp.net.core" Target="_blank">@localizer["Facebook Group"]</MudMenuItem>
+                        <MudMenuItem Link="https://www.facebook.com/codewithmukesh" Target="_blank">@localizer["Facebook Page"]</MudMenuItem>
                         <MudMenuItem Link="https://www.linkedin.com/in/iammukeshm/" Target="_blank">LinkedIn</MudMenuItem>
-                        <MudText Typo="Typo.body2" Class="px-4 py-2"><b>Resources</b></MudText>
-                        <MudMenuItem Link="https://mudblazor.com/" Target="_blank">MudBlazor Documentation</MudMenuItem>
-                        <MudMenuItem Link="https://github.com/blazorhero/CleanArchitecture" Target="_blank">Quick-Start Guide</MudMenuItem>
+                        <MudText Typo="Typo.body2" Class="px-4 py-2"><b>@localizer["Resources"]</b></MudText>
+                        <MudMenuItem Link="https://mudblazor.com/" Target="_blank">@localizer["MudBlazor Documentation"]</MudMenuItem>
+                        <MudMenuItem Link="https://github.com/blazorhero/CleanArchitecture" Target="_blank">@localizer["Quick-Start Guide"]</MudMenuItem>
                     </MudMenu>
                     <LanguageSelector />
-                    <MudTooltip Text="Toggle Dark Mode">
+                    <MudTooltip Text="@localizer["Toggle Dark Mode"]">
                         <MudIconButton Icon="@Icons.Material.Filled.Brightness4" Color="Color.Inherit" OnClick="@((e) => DarkMode())" />
                     </MudTooltip>
                     <MudTooltip Text="Fork Repository">
@@ -62,8 +63,8 @@
                     <ChildContent>
                         <UserCard Class="mt-n2" />
                         <MudDivider Class="mb-2" />
-                        <MudListItem Text="Account" Href="/account" Icon="@Icons.Material.Outlined.Person" />
-                        <MudListItem Text="Logout" Icon="@Icons.Material.Outlined.Login" OnClick="Logout" />
+                        <MudListItem Text="@localizer["Account"]" Href="/account" Icon="@Icons.Material.Outlined.Person" />
+                        <MudListItem Text="@localizer["Logout"]" Icon="@Icons.Material.Filled.Logout" OnClick="Logout" />
                     </ChildContent>
                 </MudMenu>
             </MudAppBar>
@@ -111,9 +112,11 @@
     }
     void Logout()
     {
+        string logoutConfirmationText = localizer["Logout Confirmation"];
+        string logoutText = localizer["Logout"];
         var parameters = new DialogParameters();
-        parameters.Add("ContentText", "Do you really want to Logout?");
-        parameters.Add("ButtonText", "Logout");
+        parameters.Add("ContentText", logoutConfirmationText);
+        parameters.Add("ButtonText", logoutText);
         parameters.Add("Color", Color.Error);
 
         var options = new DialogOptions() { CloseButton = true, MaxWidth = MaxWidth.Small, FullWidth = true };
